+++ packages/extensions/modelerfour/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @autorest/modelerfour
 
-This log was last generated on Thu, 07 Apr 2022 20:47:30 GMT and should not be manually modified.
+This log was last generated on Tue, 03 May 2022 20:20:34 GMT and should not be manually modified.
+
+## 4.23.3
+Tue, 03 May 2022 20:20:34 GMT
+
+### Patches
+
+- **Fix** issue when using unknown media types with a known one. Ignore those.
 
 ## 4.23.2
 Thu, 07 Apr 2022 20:47:30 GMT